+++ docker-compose.yml
@@ -174,6 +174,8 @@ services:
       - GCP_BUCKET_NAME=activitypub
       - GCP_STORAGE_EMULATOR_HOST=http://fake-gcs:4443
       - TAGS
+      - URL_GHOST_ACTIVITY_PUB=http://fake-ghost-activitypub.test
+      - URL_EXTERNAL_ACTIVITY_PUB=http://fake-external-activitypub.test
     command: /opt/activitypub/node_modules/.bin/cucumber-js
     depends_on:
       fake-ghost-activitypub:
