+++ features/support/utils.js
@@ -1,0 +1,3 @@
+export async function wait(n) {
+    return new Promise((resolve) => setTimeout(resolve, n));
+}
